test: move stage2 discount test to proper file
